Sent auxiliary functionalities to subpackaged
