+++ cvpack/helix_angle_content.py
@@ -13,7 +13,8 @@ import openmm
 from openmm import app as mmapp
 from openmm import unit as mmunit
 
-from .cvpack import BaseCollectiveVariable, SerializableResidue
+from .cvpack import BaseCollectiveVariable
+from .serializer import SerializableResidue
 from .units import ScalarQuantity
