+++ cvpack/residue_coordination.py
@@ -13,7 +13,8 @@ import openmm
 from openmm import unit as mmunit
 from openmm.app.topology import Residue
 
-from .cvpack import BaseCollectiveVariable, SerializableResidue
+from .cvpack import BaseCollectiveVariable
+from .serializer import SerializableResidue
 from .units import ScalarQuantity, value_in_md_units
